+++ src/test/sighash_tests.cpp
@@ -3,6 +3,8 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "data/sighash.json.h"
+
+#include "consensus/tx_verify.h"
 #include "main.h"
 #include "test_random.h"
 #include "serialize.h"
