+++ src/wallet/wallet.h
@@ -11,6 +11,7 @@
 
 #include "amount.h"
 #include "base58.h"
+#include "consensus/tx_verify.h"
 #include "crypter.h"
 #include "kernel.h"
 #include "key.h"
